server/requirements.txt
fastapi==0.108.0
uvicorn==0.25.0
pydantic==2.5.3
pydantic-settings==2.1.0
python-dotenv==1.0.0
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
alembic==1.13.1
python-jose==3.3.0
passlib==1.7.4
python-multipart==0.0.6
bcrypt==4.0.1
email-validator==2.1.0
pyotp==2.9.0
qrcode==7.4.2
Pillow==10.2.0
html2text==2024.2.26
jinja2==3.1.2
starlette==0.31.1 